requirements.txt
pyrogram
tgcrypto
motor
pymongo
python-dotenv
fpdf
